scanner_worker/requirements.txt
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
GitPython==3.1.40
requests==2.31.0
bandit==1.7.5
pbr==6.0.0
psycopg2-binary==2.9.11
